Address review: clean_rate_pct, f64 ring buffer, drop dead route edit
Three fixes from review feedback:

1. Replace `hydration_success_rate_pct` (hydrated / attempts, which
   mixed per-section dedup with per-chunk denominators) with
   `clean_rate_pct` — `(chunks_in − total_fallbacks) / chunks_in`.
   Trends toward 100 % when the pipeline is healthy. Tile label
   updated from "Hydration success" to "Clean rate"; info-modal copy
   reworked to explain the dedup interaction.

2. Use f64 for `gap` / `threshold` everywhere (ring buffer entries,
   averages, recorder signatures, frontend struct). The routing logic
   computes them as f64; downcasting to f32 was lossless for display
   but inconsistent.

3. Stop touching the dead `configure_agentic_monitor_routes` — it's
   defined but unreachable in production (no `App::configure` caller),
   so any edits there are no-ops. Live route registration stays in
   `api/mod.rs`. Drive-by: add the missing `/agents/rig-stats` line
   to `api/mod.rs` so the existing Rig panel stops 404-ing silently.

+++ backend/src/agent.rs
@@ -982,8 +982,8 @@ impl<'a> Agent<'a> {
                         crate::monitoring::pointer_stats::record_pointer_hydration(
                             used_chunks.len(),
                             &h,
-                            frag_gap as f32,
-                            gap_threshold as f32,
+                            frag_gap,
+                            gap_threshold,
                         );
                         tracing::info!(
                             chunks = used_chunks.len(),
